Add __pycache__ to gitignore just-in-case.
Python will store the pre-compiled binary files of all local imported objects. If this is not desired then the sys.dont_write_bytecode = True flag needs to be set within the code AND all local files need to be imported after this flag. OR an environment variable needs to be exported.  export PYTHONDONTWRITEBYTECODE=1